Refactor league and produce better documentation
